pyproject.toml
[tool.poetry]
name = "muscatplus_indexer"
version = "v1.0.4"
description = ""
authors = ["Andrew Hankinson <andrew.hankinson@rism.digital>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
dbutils = "^3.0"
mysqlclient = "^2.2.4"
pyyaml = "^6.0.0"
edtf = {git = "https://github.com/rism-digital/python-edtf.git", rev = "performance-enhancements"}
httpx = "^0.27.0"
verovio = { version = "^4.1.0", allow-prereleases = true }
sentry-sdk = "^2.0.0"
orjson = "^3.10.3"
pymarc = "^5.2.0"
#pymarc = {git = "https://gitlab.com/ahankinson/pymarc", rev = "fix-tag-length"}
psycopg = {extras = ["pool"], version = "^3.2.1"}

[tool.poetry.dev-dependencies]
ipython = "^8.0.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
priority = "supplemental"


[tool.poetry.group.dev.dependencies]
deptry = "^0.13.0"

[tool.black]
line-length = 121

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.masonry.api"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # pycodestyle warnings
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    ## Flake8 plugins
    "UP",
    # flake8-bugbear
    "B",
    # flake8-comprehensions
    "C",
    # flake8-django
    "DJ",
    # flake8-bandit
    "S",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

ignore = [
    # Ignore Pycodestyle line-length warnings, (mainly long comments).
    "E501",
    # Ignore McCabe complexity (for now).
    "C901",
]